+++ build_runner/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.8.5
+
+- If a build appears to be not making progress (no actions completed for x
+  amount of time), then a warning will now be logged with the pending actions.
+
 ## 0.8.4
 
 - Log the number of completed actions on successful builds.
@@ -17,10 +22,12 @@
   different.
 
 ## 0.8.2+2
+
 - The `.packages` file is now always created in the root of the output directory
   instead of under each top level directory.
 
 ## 0.8.2+1
+
 - Bug Fix: Correctly parse Window's paths with new `--output` semantics.
 
 ## 0.8.2